NeighborlyAPI/requirements.txt
azure-functions==1.9.0
pymongo==4.1.1
requests==2.32.3
